[FEATURE]added-role-management

+++ todo-api/Controllers/AssignmentsDeleteController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using todo_api.Data.Repos;
 using todo_api.Models;
@@ -8,17 +9,21 @@ namespace todo_api.Controllers
 {
     [Route(Constants.API_PATH + Constants.USERS_PATH + Constants.USER_ID + "/" + Constants.ASSIGNMENTS_PATH)]
     [ApiController]
-    public class AssignmentsDeleteController: ControllerBase
+    [Authorize(Roles = Constants.ADMIN_ROLE + ", " + Constants.USER_ROLE)]
+    public class AssignmentsDeleteController : ControllerBase
     {
         private readonly GetUserAssignmentUC _getUserAssignmentUC;
         private readonly DeleteAssignmentUC _deleteAssignmentUC;
+        private readonly VerifyAssignmentCreatorOrAdminUC _verifyAssignmentCreatorOrAdminUC;
 
         public AssignmentsDeleteController(
             GetUserAssignmentUC getUserAssignmentUC,
-            DeleteAssignmentUC deleteAssignmentUC)
+            DeleteAssignmentUC deleteAssignmentUC,
+            VerifyAssignmentCreatorOrAdminUC verifyAssignmentCreatorOrAdminUC)
         {
             _getUserAssignmentUC = getUserAssignmentUC;
             _deleteAssignmentUC = deleteAssignmentUC;
+            _verifyAssignmentCreatorOrAdminUC = verifyAssignmentCreatorOrAdminUC;
         }
 
         [HttpDelete(Constants.ASSIGNMENT_ID)]
@@ -28,6 +33,8 @@ namespace todo_api.Controllers
 
             if(foundAssignment == null)
                 return BadRequest(Constants.ERROR_ASSIGNMENT_NOT_FOUND);
+            if(!_verifyAssignmentCreatorOrAdminUC.VerifyAssignmentCreatorOrAdmin(User.Claims, foundAssignment.UserId))
+                return BadRequest(Constants.ERROR_NOT_ASSIGNMENT_CREATOR);
 
             await _deleteAssignmentUC.DeleteAssignment(foundAssignment);
             return NoContent();
